refactor(buildSchema): extract buildSchema in parsers-commons
